+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## UNRELEASED (0.19.0)
+
+✨ **New features**
+
+- New rounding mode: `RoundingMode::HalfOdd`
+
 ## [0.18.0](https://github.com/brick/math/releases/tag/0.18.0) - 2026-06-14
 
 💥 **Breaking changes**